+++ libft/libft.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   libft.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jcros <jcros@student.42.fr>                +#+  +:+       +#+        */
+/*   By: juliencros <juliencros@student.42.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/07 12:42:13 by jcros             #+#    #+#             */
-/*   Updated: 2022/11/08 13:17:45 by jcros            ###   ########.fr       */
+/*   Updated: 2022/11/08 15:38:08 by juliencros       ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ int ft_isascii(int i);
 int ft_isdigit (int i);
 int ft_isprint (int i);
 char *ft_memcpy (void *dest, const void *src, size_t n);
-int ft_strlen (char *str);
+size_t ft_strlen (char *str);
 void *ft_memmove (void *dest, const void *src, size_t n);
 size_t ft_strlcpy(char *dst, const char *src, size_t size);
 size_t ft_strlcat(char *dst, const char *src, size_t size);
